v4.1.0.0 - rebase
Now that I know what I'm doing, time to completely tear it up and redo everything!

I wanted to do this for a while, my motivation for the project died once I hated how inconsistent it got and it was difficult to add the punishments section.

FOR NOW: deleted everything..